requirements.txt
fastapi==0.115.2
pydantic~=2.7.3
python-dotenv==1.0.1
pytz==2024.1
redis~=4.3.4
shortuuid==1.0.12
